\name{COOLBEANS-package}
\alias{COOLBEANS-package}
\alias{COOLBEANS}
\docType{package}
\title{
\packageTitle{COOLBEANS}
}
\description{
\packageDescription{COOLBEANS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{COOLBEANS}
\packageIndices{COOLBEANS}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{COOLBEANS}

Maintainer: \packageMaintainer{COOLBEANS}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from ~~
~~ file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
