% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossvalidation_eval.R
\name{crossvalidation_eval}
\alias{crossvalidation_eval}
\title{Model evaluation after cross-validation}
\usage{
crossvalidation_eval(model, test_data, type = "regression")
}
\arguments{
\item{model}{trained model to be used}

\item{test_data}{data to be used for testing}

\item{type}{problem type either regression or classification}
}
\value{
List that includes model evaluation and some metrics
}
\description{
The function trains a model using linear regression (elastic net)
and cross-validation
}
