% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossvalidation_model.R
\name{crossvalidation_model}
\alias{crossvalidation_model}
\title{Model training using cross-validation and the model of choice}
\usage{
crossvalidation_model(
  train_data,
  model = "glmnet",
  nfolds = 2,
  nrepeats = 2,
  ltune = 5
)
}
\arguments{
\item{train_data}{data to be used for training}

\item{model}{model to be used, default is elastic model}

\item{nfolds}{number of folds to be used}

\item{nrepeats}{number of repeats to be used}

\item{ltune}{tuneLength parameter during training}
}
\value{
Output from training model
}
\description{
The function trains a model using linear regression (elastic net)
and cross-validation
}
