% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Standard preprocessing}
\usage{
preprocess_data(raw_data, cutoff_met = 0, cutoff_subj = 0)
}
\arguments{
\item{raw_data}{dataframe with metabolomic data}
}
\value{
dataframe
}
\description{
Standard preprocessing
}
