% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocessing}
\alias{preprocessing}
\title{Standard data wrangling (pre-processing)}
\usage{
preprocessing(
  raw_data,
  id,
  target,
  start_metabolites,
  cutoff_columns = 0.2,
  cutoff_rows = 0.2,
  imputation = "median"
)
}
\arguments{
\item{raw_data}{dataframe with metabolomic data}

\item{id}{id of each row}

\item{target}{target to focus in the model}

\item{start_metabolites}{column number where metabolites start}

\item{cutoff_columns}{threshold to use to discard columns given n NaN}

\item{cutoff_rows}{threshold to use to discard rows given n NaN}

\item{imputation}{method of choice, four available: median, +mean, lower, knn}
}
\value{
A [tibble::tibble()]. It changes name columns for column ID
and column target.
}
\description{
This function performs a simple workflow of data wrangling.
Overall it will remove noise data, duplicate rows and columns/rows with NAs.
Next the data will be transformed using imputation, log transformation and
normalization. The function will also assign new column names to ID column
and target column which will be used later for the other functions.
}
