% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_calculation.R
\name{score_calculation}
\alias{score_calculation}
\title{Calculation of scores using residuals from lm and coefficients from the CV model}
\usage{
score_calculation(ids, df, coefficients)
}
\arguments{
\item{ids}{calculated from lm}

\item{df}{data to calculate score}

\item{coefficients}{calculated from ml with crossvalidation}
}
\value{
data frame with scores per sample
}
\description{
The function calculates a score based on residuals and coefficients
}
