% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{splitting}
\alias{splitting}
\title{Splits the prepossessed data set into train and test}
\usage{
splitting(preprocessed_data, target, training_split = 0.7)
}
\arguments{
\item{preprocessed_data}{data after pre-processing}

\item{target}{column}

\item{training_split}{data fraction to use for training}
}
\value{
A list of two [tibble::tibble()] for training and testing dataset.
}
\description{
The function splits the data for model training, two datasets
are generated for training and testing. It is possible to define the percentage
of data wanted for training, default = 0.7.
}
