% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_model.R
\name{test_model}
\alias{test_model}
\title{Title}
\usage{
test_model(model, test_data)
}
\arguments{
\item{model}{to be use with the data}

\item{test_data}{data to test the model}
}
\value{
model coefficients
}
\description{
Title
}
