% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_model_class_knn.R
\name{testing_knn_classification}
\alias{testing_knn_classification}
\title{Model testing from k-nearest neighbor model using classification algorithm}
\usage{
testing_knn_classification(model, test_data)
}
\arguments{
\item{model}{to be use with the data}

\item{test_data}{data to test the model}
}
\value{
A list of one [tibble::tibble()] and one numeric value
}
\description{
The function asses the performance of the trained model using
k-nearest neighbors with classification algorithm.
}
