% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_model_regression_knn.R
\name{testing_knn_regression}
\alias{testing_knn_regression}
\title{Model testing from k-nearest neighbor model using regression algorithm}
\usage{
testing_knn_regression(model, test_data)
}
\arguments{
\item{model}{to be use with the data}

\item{test_data}{data to test the model}
}
\value{
list of one [tibble::tibble()] and two numeric values
}
\description{
The function asses the performance of the trained model using
random forest with regression algorithm.
}
