% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_model.R
\name{testing_lr}
\alias{testing_lr}
\title{Model testing from linear regression model}
\usage{
testing_lr(model, test_data)
}
\arguments{
\item{model}{to be use with the data}

\item{test_data}{data to test the model}
}
\value{
A list of one [tibble::tibble()] and one numeric value
}
\description{
The function asses the performance of the trained model using
linear regression.
}
