% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_model_regression.R
\name{testing_rf_regression}
\alias{testing_rf_regression}
\title{Model testing from random forest model using regression algorithm}
\usage{
testing_rf_regression(model, test_data)
}
\arguments{
\item{model}{to be use with the data}

\item{test_data}{data to test the model}
}
\value{
A list of two [tibble::tibble()] and two numeric values
}
\description{
The function asses the performance of the trained model using
random forest with regression algorithm.
}
