% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model_knn.R
\name{training_knn}
\alias{training_knn}
\title{Model training using k-nearest neighbor}
\usage{
training_knn(train_data, type = 1)
}
\arguments{
\item{train_data}{data to be used for training}

\item{type}{two options 1 for regression using linear regression and 2 for classification using random forest.}
}
\value{
A list of two elements a list and an stage class
}
\description{
The function trains a model using k-nearest neighbor. It is possible
to use two different algorithms: regression or classification. The number of clusters
will be defined based on the unique elements on the target column up to n = 10.
}
