% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model.R
\name{training_lr}
\alias{training_lr}
\title{Model training using linear regression}
\usage{
training_lr(train_data)
}
\arguments{
\item{train_data}{data to be used for training}
}
\value{
A list of two elements a list and an stage class
}
\description{
The function trains a model using linear regression (elastic net)
}
