% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model_rf.R
\name{training_rf}
\alias{training_rf}
\title{Model training using random forest}
\usage{
training_rf(train_data, type = 1)
}
\arguments{
\item{train_data}{data to be used for training}

\item{type}{two options 1 for regression using linear regression and 2 for classification using random forest.}
}
\value{
A list of two elements a list and an stage class
}
\description{
The function trains a model using random forest. It is possible
to use two different algorithms: regression or classification.
}
