% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sing_met.R
\name{sing_met_analysis}
\alias{sing_met_analysis}
\title{Analysis of single metabolites based on a target (and confounder/s)}
\usage{
sing_met_analysis(
  data,
  exposure_feature,
  start_metabolites,
  confounders = NULL,
  threshold = 0.1,
  correction = NULL
)
}
\arguments{
\item{data}{data frame to use for analysis}

\item{exposure_feature}{variable of exposure name (e.g. diet score, group of exposure)}

\item{start_metabolites}{column number where metabolites start}

\item{confounders}{list of confounders}

\item{correction}{correction method to use}
}
\value{
A [tibble::tibble()].
}
\description{
The function fits a linear model where the response variable are
each metabolite and the terms are the exposure feature and other features that
are defined as confounders.
}
