% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{split_data}
\alias{split_data}
\title{Split dataset}
\usage{
split_data(preprocessed_data, training_split = 0.7)
}
\arguments{
\item{training_split}{}
}
\value{
split dataframes
}
\description{
Split dataset
}
