% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{splitting}
\alias{splitting}
\title{Splits the data set after the data wrangling steps and feature selection into train and test}
\usage{
splitting(clean_data, target, training_split = 0.7)
}
\arguments{
\item{clean_data}{data after data wrangling and feature selection}

\item{target}{column}

\item{training_split}{data fraction to use for training}
}
\value{
A list of two [tibble::tibble()] for training and testing data set.
}
\description{
The function splits the data for model training, two datasets
are generated for training and testing. It is possible to define the percentage
of data wanted for training, default = 0.7.
}
